* commit '9453fcde8ac4e8a58a74b73c49798608991d7f30':
  REG-7553: Refaktorering
  REG-7553: Returnerer status ved oppdatering av eksisterende arbeidsforhold
  REG-7553: La til test av mapping
  REG-7553: Oppdaterer permisjon, utenlandsopphold og antallTimerForTimelønnet. Endret antall timer fra Integer til Double
  REG-7553: Passer på at arbeidsforhold ikke er null under håndtering
  REG-7553: La til test. Finner seneste arbeidsavtale under oppdatering
  REG-7553: Sonar
  REG-7553: Setter aktørtype til arbeidstaker og arbeidsgiver under oppdatering
  REG-7553: Sender oppdateringer til aareg. Fikset tester.
  REG-7553: Utkast til ny månedlig oppdatering av arbeidsforhold

+++ aareg-domain/src/main/java/no/nav/registre/aareg/domain/RsAntallTimerForTimeloennet.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 @Builder
 public class RsAntallTimerForTimeloennet {
 
-    private Integer antallTimer;
+    private Double antallTimer;
     private RsPeriode periode;
 }
